Remove requirement from entity integration (#30113)
