+++ homeassistant/components/doods/manifest.json
@@ -1,10 +1,8 @@
 {
-    "domain": "doods",
-    "name": "DOODS - Distributed Outside Object Detection Service",
-    "documentation": "https://www.home-assistant.io/integrations/doods",
-    "requirements": [
-        "pydoods==1.0.2"
-    ],
-    "dependencies": [],
-    "codeowners": []
+  "domain": "doods",
+  "name": "DOODS - Distributed Outside Object Detection Service",
+  "documentation": "https://www.home-assistant.io/integrations/doods",
+  "requirements": ["pydoods==1.0.2", "pillow==6.2.1"],
+  "dependencies": [],
+  "codeowners": []
 }
