+++ homeassistant/components/image_processing/manifest.json
@@ -2,11 +2,7 @@
   "domain": "image_processing",
   "name": "Image processing",
   "documentation": "https://www.home-assistant.io/integrations/image_processing",
-  "requirements": [
-    "pillow==6.2.1"
-  ],
-  "dependencies": [
-    "camera"
-  ],
+  "requirements": [],
+  "dependencies": ["camera"],
   "codeowners": []
 }
