+++ homeassistant/components/tensorflow/manifest.json
@@ -2,7 +2,12 @@
   "domain": "tensorflow",
   "name": "Tensorflow",
   "documentation": "https://www.home-assistant.io/integrations/tensorflow",
-  "requirements": ["tensorflow==1.13.2", "numpy==1.17.4", "protobuf==3.6.1"],
+  "requirements": [
+    "tensorflow==1.13.2",
+    "numpy==1.17.4",
+    "protobuf==3.6.1",
+    "pillow==6.2.1"
+  ],
   "dependencies": [],
   "codeowners": []
 }
