+++ requirements_test_all.txt
@@ -320,11 +320,6 @@ pexpect==4.6.0
 # homeassistant.components.pilight
 pilight==0.1.1
 
-# homeassistant.components.image_processing
-# homeassistant.components.proxy
-# homeassistant.components.qrcode
-pillow==6.2.1
-
 # homeassistant.components.plex
 plexapi==3.3.0
